Phase 5 commit 5: extract LinkPickerDropdown.tsx; relocate LiveEditor
- editor/LinkPickerDropdown.tsx: forwardRef-style dropdown component
  with useImperativeHandle keyboard navigation. The hooks-extracted
  picker state (useLinkPicker) and the dropdown component are now
  cleanly separated.
- LiveEditor.tsx moved to notes/editor/LiveEditor.tsx alongside
  the markdown helpers, upload, and dropdown it depends on. Imports
  in Notes.tsx updated.

LiveEditor.tsx 301 → 247 lines (was 696 before Phase 5). Build clean.
